refactor(theme): provide proper layout for 404 page
