+++ packages/gatsby-theme-docs/src/layouts/not-found.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import { useSiteData } from '../hooks/use-site-data';
+import LayoutApplication from './internals/layout-application';
+import LayoutCentered from './internals/layout-centered';
+import LayoutHeader from './internals/layout-header';
+import LayoutPage from './internals/layout-page';
+import LayoutFooter from './internals/layout-footer';
+import LayoutMain from './internals/layout-main';
+import LayoutPageContent from './internals/layout-page-content';
+import PageContentInset from './internals/page-content-inset';
+
+const PageNotFound = props => {
+  const siteData = useSiteData();
+  return (
+    <LayoutApplication>
+      <LayoutHeader siteTitle={siteData.siteMetadata.title} />
+      <LayoutCentered>
+        <LayoutMain>
+          <LayoutPage>
+            <LayoutPageContent>
+              <PageContentInset>
+                {
+                  // eslint-disable-next-line react/prop-types
+                  props.children
+                }
+              </PageContentInset>
+              <LayoutFooter />
+            </LayoutPageContent>
+          </LayoutPage>
+        </LayoutMain>
+      </LayoutCentered>
+    </LayoutApplication>
+  );
+};
+
+export default PageNotFound;
